pyproject.toml
[project]
name = "workua-resume-toolkit"
version = "0.1.0"
description = "A small toolkit to scrape Work.ua resumes, parse them, and store/inspect results."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name="Danylo Dvoichenkov"}]
dependencies = [
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.2",
  "lxml>=5.0.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
db = ["psycopg2-binary>=2.9.9"]
ui = ["streamlit>=1.31.0"]
semantic = [
  "sentence-transformers>=2.2.2",
  "faiss-cpu>=1.7.4",
  "numpy>=1.24.0",
]

[project.scripts]
workua-init-db = "workua_toolkit.scripts.init_db:main"
workua-toolkit = "workua_toolkit.cli.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
